discord-wowaudit-bot/discord-wowaudit-bot/requirements.txt
discord.py>=2.3
python-dotenv>=1.0
requests>=2.31
